+++ Runtime/Scripts/DebugItem/DebugButton.cs
@@ -1,8 +1,7 @@
 using System;
-using System.Reflection.Emit;
 using UnityEngine;
 using UnityEngine.UI;
-namespace DevPanel
+namespace DeveloperMenu.DebugItems
 {
     public class DebugButton : DebugItem
     {
@@ -10,7 +9,7 @@ namespace DevPanel
 
         private event Action action;
 
-        public void Initialize(Header header, Action debugAction, Settings s = new())
+        public void Initialize(Header header, Action debugAction)
         {
             _header = header;
 
@@ -18,10 +17,8 @@ namespace DevPanel
             label.text = header.name;
 
             action = debugAction;
-            _settings = s;
 
             button.onClick.AddListener(action.Invoke);
-            if(s.closeDebugOnClick) button.onClick.AddListener(()=>DebugPanel.Instance.SetMenuActive(false));
         }
     }