+++ Runtime/Scripts/DebugItem/DebugTab.cs
@@ -1,7 +1,8 @@
+using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace DevPanel
+namespace DeveloperMenu.DebugItems
 {
     public class DebugTab : DebugItem
     {
@@ -12,21 +13,32 @@ namespace DevPanel
 
         public override Transform GetItemContainer() => itemContainer.transform;
 
-        public void Initialize(Header h, Settings s)
+        public void Initialize(Header h)
         {
             _header = h;
-            _settings = s;
 
             label.text = h.name;
             name = h.name;
         }
 
+        /// Called by TextContent Button
         public void ToggleTab()
         {
             itemContainer.SetActive(!itemContainer.activeSelf);
 
             plusMinusText.text = itemContainer.activeSelf ? "-" : "+";
 
+        }
+
+        public async void OnSubItemDestroyed()
+        {
+            await Task.Delay(1); //Delay required to properly recognize the change in childcount
+
+            Debug.Log($"{header.name}: Sub Item Destroyed, remaining subitems: {GetItemContainer().childCount}", this);
+            if (GetItemContainer().childCount == 0)
+            {
+                RequestDestroy();
+            }
         }
     }