+++ Runtime/Scripts/DebugItemReference.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
+using DeveloperMenu.DebugItems;
 
-namespace DevPanel
+namespace DeveloperMenu
 {
     /// <summary>
     /// DebugBehavior acts as a reference count to the debug panel. It should be instantiated using DebugPanel.Instance.CreateDebugEvent.
@@ -8,13 +9,12 @@ namespace DevPanel
     public class DebugItemReference : MonoBehaviour
     {
 
-        [SerializeReference] private DebugItem _item;
+        private DebugItem _item;
         public DebugItem Item => _item;
 
         public void Initialize(DebugItem i)
         {
             _item = i;
-            if (!Item.settings.showDisabled) Item.gameObject.SetActive(gameObject.activeInHierarchy);
         }
 
         private void OnEnable()
@@ -28,14 +28,14 @@ namespace DevPanel
         {
             if (Item == null) return;
 
-            if (!Item.settings.showDisabled) Item.gameObject.SetActive(false);
+            Item.gameObject.SetActive(false);
         }
 
         private void OnDestroy()
         {
             if (Item != null)
             {
-                Destroy(Item.gameObject);
+                Item.RequestDestroy();
             }
         }
     }